1 Corrigindo nomes de variáveis que não fazia sentido (retirando várias referências a tuples, quando podiam não ser de fato tuples) 2 Corrigindo cast que estava para tuple no CompareSimpleRegister 3 Possibilitando o update para o listener de Folders

+++ trunk/project/datawatcher/src/main/java/br/com/datawatcher/entity/SimpleRegister.java
@@ -9,6 +9,6 @@ package br.com.datawatcher.entity;
  */
 public abstract class SimpleRegister {
 
-	public abstract int hashSimpleRegister();
+	public abstract long hashSimpleRegister();
 	
 }
